requirements.txt
flask
requests
python-dotenv
schedule